pyproject.toml
[tool.poetry]
name = "aiogram-starter-kit"
version = "0.1.0"
description = "Aiogram Starter Kit."
authors = [
  "Serhii Ihnatiev <141933729+iamihnatiev@users.noreply.github.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
aiogram = "^3.0.0"
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
